backend/requirements.txt
Flask>=2.0
flask-cors
tensorflow>=2.11
numpy
pillow
requests
beautifulsoup4
gymnasium
stable-baselines3
